% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{centerTitle}
\alias{centerTitle}
\title{Add a centered title to a plot}
\usage{
centerTitle(p, title, ...)
}
\arguments{
\item{p}{A ggplot object.}

\item{title}{Plot title.}

\item{...}{Other arguments passed to \code{ggplot2::element_text}.}
}
\value{
A ggplot object.
}
\description{
This function adds a centered title to a ggplot object.
}
