% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_plot.R
\name{circleCoords}
\alias{circleCoords}
\title{Store the radii of the circles and the corresponding number of edges}
\usage{
circleCoords(itemCoordsDF, extraCircles = 0)
}
\arguments{
\item{itemCoordsDF}{Dataframe wih item coordinates}

\item{extraCircles}{Number of circles drawn beyond those required to include
the points representing the genes.}
}
\value{
A data frame containing the radius and the number of edges for each
circle.
}
\description{
This function stores the radii of the circles and the corresponding number
of edges.
}
\keyword{internal}
