% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_plot.R
\name{densityPlot}
\alias{densityPlot}
\title{Create density plot}
\usage{
densityPlot(
  df,
  title = "Density plot",
  colorScheme = c("sea", "lava", "custom"),
  drawNN = TRUE,
  palette = NULL,
  segColor = "sandybrown",
  pointSize = 1,
  segType = c("solid", "dashed", "dotted", "dotdash", "longdash", "twodash"),
  segWidth = 0.5,
  legendPos = "right",
  nGridPoints = 300,
  expandPerc = 20,
  labelSize = 3,
  labelColor = "black",
  labelRepulsion = 1,
  labelPull = 1,
  maxOverlaps = Inf,
  ...
)
}
\arguments{
\item{df}{A data frame with at least two columns, representing the \code{x}
and \code{y} coordinates of the points. A score column can also be provided
as the third column. Nearest neighbor information can be provided in the last
column as a character vector with elements selected from the rownames.}

\item{title}{Plot title.}

\item{colorScheme}{Color scheme.}

\item{drawNN}{Whether to draw segments linking each point to its nearest
neighbor.}

\item{palette}{Color palette. Used only if color scheme is set to 'custom'.}

\item{segColor}{Nearest neighbor segment color. Ignored if \code{drawNN} is
set to \code{FALSE}.}

\item{pointSize}{Point size.}

\item{segType}{Nearest neighbor segment type. Must choose between 'solid',
'dashed', 'dotted','dotdash', 'longdash' and 'twodash'. Ignored if
\code{drawNN} is set to \code{FALSE}.}

\item{segWidth}{Nearest neighbor segment width. Ignored if \code{drawNN} is
set to \code{FALSE}.}

\item{legendPos}{Legend position.}

\item{expandPerc}{Percentage by which the grid will expanded.}

\item{labelSize}{Size of labels of strata elements.}

\item{labelColor}{Label colors.}

\item{labelRepulsion}{Repulsion strength between labels.}

\item{labelPull}{Attraction strength between a text label
and its data point.}

\item{maxOverlaps}{Maximum overlaps.}

\item{...}{Other arguments passed to centerTitle.}

\item{Number}{of grid points in each direction.}
}
\value{
A ggplot object.
}
\description{
This function creates a density plot.
}
\examples{
x <- c(1, 2, 3, 4, 6, 7, 8, 10, 12, 11, 3, 6, 4, 1, 13, 13, 14, 18, 16)
y <- c(1, 3, 1, 4, 3, 2, 8, 2, 1, 11, 8, 8, 10, 14, 13, 11, 11, 12,15)
z <- round(runif(19, 75, 100), 2)
df <- data.frame(x, y, z)
rownames(df) <- paste0('p', rownames(df))
densityPlot(df)

}
