% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dfEuclidean}
\alias{dfEuclidean}
\title{Find the Euclidean distance between two points in a data frame}
\usage{
dfEuclidean(df, i, j, pointCoords = c(1, 2), replaceZero = 0)
}
\arguments{
\item{df}{A data frame containing the coordinates
of the points.}

\item{i}{Row name or index.}

\item{j}{Row name or index}

\item{pointCoords}{The row names or indices of point coordinates.}

\item{replaceZero}{Value to replace zero with. If set to 0, no replacement
will occur.}
}
\value{
A numeric value.
}
\description{
This function finds the Euclidean distance between two points in a data
frame.
}
\keyword{internal}
