% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dfPoint}
\alias{dfPoint}
\title{Extract point from data frame of points.}
\usage{
dfPoint(df, i, pointCoords = c(1, 2))
}
\arguments{
\item{df}{A data frame containing the coordinates
of the points.}

\item{i}{Row name or index.}

\item{pointCoords}{The row names or indices of point coordinates.}
}
\value{
A numeric vector containing the point coordinates.
}
\description{
This function extracts a point from a data frame of points.
}
\keyword{internal}
