% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{dpColors}
\alias{dpColors}
\title{Create the default 'sea' palette used by \code{densityPlot}}
\usage{
dpColors()
}
\value{
A character vector of colors.
}
\description{
This function returns the 'sea' palette used by \code{densityPlot}.
}
