% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{hpColors}
\alias{hpColors}
\title{Create the default hullPlot palette}
\usage{
hpColors()
}
\value{
A character vector of colors.
}
\description{
This function returns the default palette used by \code{hullPlot}.
}
