% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hull_plot.R
\name{hullPlot}
\alias{hullPlot}
\title{Plot the convex hull of a set of points}
\usage{
hullPlot(
  pointsDF,
  title = "Hull plot",
  xInt = NULL,
  yInt = NULL,
  borderColor = "navy",
  palette = hpColors(),
  showHull = FALSE,
  xLab = "x",
  yLab = "y",
  legendLabs = paste0("Group ", seq(4)),
  legendPos = "bottom",
  pointShape = 20,
  alpha = 0.5,
  labelDF = NULL,
  labelSize = 2,
  labelColor = "black",
  maxOverlaps = 10,
  ...
)
}
\arguments{
\item{pointsDF}{A data frame with the x and y coordinates of the points.}

\item{title}{Plot title.}

\item{xInt}{The coordinate where the vertical line intersects the x axis.}

\item{yInt}{The coordinate where the horizontal line intersects the y axis.}

\item{borderColor}{The color of the horizontal and vertical dividing lines,
if provided. If \code{NULL}, no dividing lines will be drawn, though the
hull will still be split along these lines (if \code{xInt}
and/or \code{yInt}are not \code{NULL}).}

\item{palette}{Color palette.}

\item{showHull}{Whether to display the segments on the convex hull.}

\item{xLab}{x axis label.}

\item{yLab}{y axis label.}

\item{legendLabs}{Legend labels.}

\item{legendPos}{Legend position.}

\item{pointShape}{Point shape.}

\item{alpha}{Opaqueness level for the colors of the alluvia.}

\item{labelDF}{Label data frame.}

\item{labelSize}{Label size. Ignored if \code{labelDF} is \code{NULL}.}

\item{labelColor}{Label color. Ignored if \code{labelDF} is \code{NULL}.}

\item{maxOverlaps}{Maximum overlaps. Ignored if \code{labelDF}
is \code{NULL}.}

\item{...}{Other arguments passed to centerTitle.}
}
\value{
A ggplot object.
}
\description{
This function plots the convex hull of a set of points and optionally draws
a vertical and a horizontal line, dividing the hull into areas of different
colors.
}
\examples{
pointsDF <- data.frame(rank = c(1, 2, 4, 7, 10,
12, 13, 15, 16),
n = c(1, 1, 2, 3, 3, 2,
1, 2, 1))
hullPlot(pointsDF, 8.5)

}
