% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{isPointOnSeg}
\alias{isPointOnSeg}
\title{Check if a point is on a segment}
\usage{
isPointOnSeg(xPoint, yPoint, xStart, yStart, xEnd, yEnd)
}
\arguments{
\item{xPoint}{x coordinate of point P.}

\item{yPoint}{y coordinate of point P.}

\item{xStart}{x coordinate of point A.}

\item{yStart}{y coordinate of point A.}

\item{xEnd}{x coordinate of point B.}

\item{yEnd}{y coordinate of point B.}
}
\value{
Logical; whether the point is on the segment.
}
\description{
This function checks if a point P is on a segment AB.
}
\examples{
isPointOnSeg(2, 3, 1, 2, 3, 4)
isPointOnSeg(2, 3, 1, 2, 3, 8)
isPointOnSeg(4, 5, 1, 2, 3, 4)

}
