% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{labelPoints}
\alias{labelPoints}
\title{Label points in a ggplot object}
\usage{
labelPoints(p, labelDF, labelSize = 2, labelColor = "black", maxOverlaps = 10)
}
\arguments{
\item{p}{A ggplot object.}

\item{labelDF}{Label data frame.}

\item{labelSize}{Label size.}

\item{labelColor}{Label colors}

\item{maxOverlaps}{Maximum overlaps.}
}
\value{
A ggplot object.
}
\description{
This function labels points in a ggplot object
}
