% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_plot.R
\name{pointsOnCircle}
\alias{pointsOnCircle}
\title{Generate the coordinates of points on a circle centered at origin}
\usage{
pointsOnCircle(r, nPoints, seed = 50)
}
\arguments{
\item{r}{Radius.}

\item{nPoints}{Number of points.}

\item{seed}{Random seed.}
}
\value{
A data frame with the coordinates of the points.
}
\description{
This function generates nPoints on a circle of radius r
centered at origin.
}
\keyword{internal}
