% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{pointsToSegments}
\alias{pointsToSegments}
\title{Construct a data frame of segments from a data frame of points}
\usage{
pointsToSegments(pointsDF, formPolygon = TRUE)
}
\arguments{
\item{pointsDF}{A data frame with the x and y coordinates of the points.
Each point must appear only once.}

\item{formPolygon}{Whether to close the figure by adding an extra segment.}
}
\value{
A data frame of segments.
}
\description{
This function constructs a data frame of segments from a data frame of
points.
}
