% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hull_plot.R
\name{quadBorders}
\alias{quadBorders}
\title{Find the coordinates of the points establishing the four divisions of
the hull}
\usage{
quadBorders(pointsDF, xInt, yInt, vCoords, hCoords)
}
\arguments{
\item{pointsDF}{A data frame with the x and y coordinates of the points.}

\item{xInt}{The coordinate where the vertical line intersects the x axis.}

\item{yInt}{The coordinate where the horizontal line intersects the y axis.}

\item{vCoords}{The y coordinates of the two points where the vertical line
intersects the convex hull.}

\item{hCoords}{The x coordinates of the two points where the horizontal line
intersects the convex hull.}
}
\value{
A data frame with 2 columns representing the 12 points (not unique)
determining the boundaries of the hull divisions.
}
\description{
This function finds the coordinates of the points establishing the
four divisions of the hull
}
\keyword{internal}
