% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_plot.R
\name{radialPlot}
\alias{radialPlot}
\title{Draw radial plot for a degrees data frame}
\usage{
radialPlot(
  degreesDF,
  title = "Radial plot",
  degreeLegendTitle = "Degree",
  groupLegendTitle = "Group",
  extraCircles = 0,
  palette = rpColors(length(unique(degreesDF[, 3]))),
  labelSize = 3,
  pointSize = 0.8,
  labelRepulsion = 1,
  labelPull = 0,
  maxOverlaps = 10,
  ...
)
}
\arguments{
\item{degreesDF}{Degree data frame with names on the first column and
degrees on the second column.}

\item{title}{Plot title.}

\item{degreeLegendTitle}{Degree legend title.}

\item{groupLegendTitle}{Group legend title.}

\item{extraCircles}{Number of circles drawn beyond those needed to include
the points representing the genes.}

\item{palette}{Color palette.}

\item{labelSize}{Label size.}

\item{pointSize}{Point size.}

\item{labelRepulsion}{Repulsion strength between labels.}

\item{labelPull}{Attraction strength between a text label
and its data point.}

\item{maxOverlaps}{Maximum overlaps. Ignored if \code{labelDF}
is \code{NULL}.}

\item{...}{Other arguments passed to centerTitle.}
}
\value{
A ggplot object.
}
\description{
This function draws a radial plot for a degrees data frame.
}
\examples{
degreesDF <- data.frame(Protein = paste0('P', seq(20)),
Degree = sample(10, 20, replace=TRUE),
Group = sample(3, 20, replace=TRUE))
radialPlot(degreesDF)

}
