% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_plot.R
\name{rankPlot}
\alias{rankPlot}
\title{Create a rank plot}
\usage{
rankPlot(df, title = "Rank plot", viridisPal = "turbo", xLab = "Item", ...)
}
\arguments{
\item{df}{A data frame with ranks as columns, items as rows.}

\item{title}{Plot title.}

\item{viridisPal}{Viridis palette.}

\item{xLab}{Label of x axis.}

\item{...}{Other arguments passed to centerTitle.}
}
\value{
A ggplot object.
}
\description{
This function creates a rank plot
}
\examples{
df <- do.call(cbind, lapply(seq(30), function(i) sample(10, 10)))
rownames(df) <- paste('M', seq(10))
colnames(df) <- paste('R', seq(30))
rankPlot(df)

}
