% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_plot.R
\name{rankSummary}
\alias{rankSummary}
\title{Create a rank summary}
\usage{
rankSummary(df)
}
\arguments{
\item{df}{A data frame with ranks as columns, items as rows.}
}
\value{
A rank summary dataframe.
}
\description{
This function creates a summary of multiple ranks provided for input items.
}
\keyword{internal}
