% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_plot.R
\name{reorderDF}
\alias{reorderDF}
\title{Sort a data frame by the first column and convert the second to a factor}
\usage{
reorderDF(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
An object of class \code{gg}.
}
\description{
This function sort a data frame by the first column and convert the second
to a factor.
}
\examples{
df <- data.frame(a = c(2, 4, 1, 3, 6),
b = c(2, 8, 3, 19, 3))
reorderDF(df)

}
