% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{rpColors}
\alias{rpColors}
\title{Create a palette designed to represent dots over a viridis background}
\usage{
rpColors(nColors = 10)
}
\arguments{
\item{nColors}{Number of colors.}
}
\value{
A character vector of colors.
}
\description{
This function returns a 10-color palette used as the default
of \code{radialPlot}.
}
