% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{Create a volcano plot}
\usage{
volcanoPlot(
  df,
  title = NULL,
  logCol = "avg_log2FC",
  pvalCol = "p_val_adj",
  xLab = expression(log[2] ~ fold ~ change),
  yLab = expression(-log[10] ~ `p-value`),
  legendTitle = "Significance",
  legendLabs = c("Not significant", expression(log[2] ~ FC), "p-value",
    expression(`p-value` ~ and ~ log[2] ~ FC)),
  legendPos = c("right", "top", "left", "bottom"),
  logFCThr = 1,
  pvalThr = 1e-05,
  labeledGenes = NULL,
  labLogFCThr = 1.8,
  labPvalThr = 1e-12,
  labelType = c("boxed", "free"),
  labelSize = 2.5,
  labelColor = "black",
  labelRepulsion = 1,
  labelPull = 0,
  maxOverlaps = 100,
  pointSize = 1,
  alpha = 0.7,
  palette = c("dodgerblue4", "goldenrod2", "slateblue3", "red2"),
  legendTextSize = 10,
  legendTitleSize = 10,
  axisTextSize = 12,
  axisTitleSize = 12,
  theme = c("linedraw", "bw", "classic", "minimal"),
  ...
)
}
\arguments{
\item{df}{A data frame with rownames as genes, one log column and one
p-value column.}

\item{title}{Plot title.}

\item{logCol}{Log column.}

\item{pvalCol}{P-value column.}

\item{xLab}{Label of x axis.}

\item{yLab}{Label of y axis.}

\item{legendTitle}{Legend title.}

\item{legendLabs}{Legend labels.}

\item{legendPos}{Legend position.}

\item{logFCThr}{Threshold used to separate significant log values.}

\item{pvalThr}{Threshold used to separate significant p-values.}

\item{labeledGenes}{Gene labels to be displayed on the plot. Default is
\code{NULL} (gene labels will be displayed on the basis
of \code{labLogFCThr }) and \code{labPvalThr}).}

\item{labLogFCThr}{Threshold used to plot gene labels based on log values.
Ignored if \code{labeledGenes} is not \code{NULL}.}

\item{labPvalThr}{Threshold used to plot gene labels based on p-values.
Ignored if \code{labeledGenes} is not \code{NULL}.}

\item{labelType}{Whether to draw a box around labels (option 'boxed') or not
(option 'free'). Default is 'free'.}

\item{labelSize}{Label size.}

\item{labelColor}{Label color.}

\item{labelRepulsion}{Repulsion strength between labels.}

\item{labelPull}{Attraction strength between a text label
and its data point.}

\item{maxOverlaps}{Maximum overlaps.}

\item{pointSize}{Point size.}

\item{alpha}{Opaqueness level of point color.}

\item{palette}{Color palette.}

\item{legendTextSize}{Legend text size.}

\item{legendTitleSize}{Legend title size.}

\item{axisTextSize}{Axis text size.}

\item{axisTitleSize}{Axis title size.}

\item{theme}{Plot theme. Choose between 'linedraw', 'bw', 'classic' and
'minimal'. Default is 'linedraw'.}

\item{...}{Additional arguments passed to \code{centerTitle}.}
}
\value{
An object of class \code{gg}.
}
\description{
This function creates a volcano plot.
}
