% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_plot.R
\name{circleCoords}
\alias{circleCoords}
\title{Store the radii of the circles and the corresponding values}
\usage{
circleCoords(itemCoordsDF, extraCircles = 0)
}
\arguments{
\item{itemCoordsDF}{Data frame wih item coordinates.}

\item{extraCircles}{Number of circles drawn beyond those required to include
the points.}
}
\value{
A data frame containing the radius and the corresponding value for
each circle.
}
\description{
This function stores the radii of the circles and the corresponding value
for each circle.
}
\keyword{internal}
