% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{documentFun}
\alias{documentFun}
\title{Internal function used for documenting other functions}
\usage{
documentFun(
  title = NULL,
  xLab = "x",
  yLab = "y",
  legendPos = "right",
  legendLab = "Label",
  palette = "Spectral",
  viridisPal = "turbo",
  labelSize = 2.5,
  labelColor = "black",
  pointSize = 0.8,
  legendTitleSize = 10,
  legendTextSize = 10,
  axisTextSize = 12,
  axisTitleSize = 12,
  xAngle = 45,
  vJust = 0.6,
  margins = margin(0, -10, -10, -10),
  ...
)
}
\arguments{
\item{title}{Plot title.}

\item{xLab}{Label of x axis.}

\item{yLab}{Label of y axis.}

\item{legendPos}{Legend position.}

\item{legendLab}{Legend label.}

\item{palette}{Color palette.}

\item{viridisPal}{Viridis palette.}

\item{labelSize}{Label size.}

\item{labelColor}{Label color.}

\item{pointSize}{Point size.}

\item{legendTitleSize}{Legend title size.}

\item{legendTextSize}{Legend text size.}

\item{axisTextSize}{Axis text size.}

\item{axisTitleSize}{Axis title size.}

\item{xAngle}{Angle of x axis text.}

\item{vJust}{Vertical justification in [0, 1].}

\item{margins}{Plot margins. Must be a vector of size 4 listing the desired
top, right, bottom and left margin, in that order.}

\item{...}{Additional arguments passed to \code{centerTitle}.}
}
\value{
\code{NULL}. This function is only used internally for
documentation.
}
\description{
This function is used internally to help document other functions.
}
\keyword{internal}
