% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{dpColors}
\alias{dpColors}
\title{Create a palette designed for densityPlot}
\usage{
dpColors(palette = c("oasis", "sea", "lava"))
}
\arguments{
\item{palette}{One of 'sea' and 'lava'.}
}
\value{
A character vector of colors.
}
\description{
This function returns a palette designed for \code{densityPlot}.
}
