% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{dpColors}
\alias{dpColors}
\title{Create a palette designed for densityPlot}
\usage{
dpColors(
  palette = c("cake", "cloudy", "grapes", "lava", "oasis", "orichalc", "sea", "sky")
)
}
\arguments{
\item{palette}{One of 'lava', 'oasis', 'orichalc', 'sea' and 'sky'.}
}
\value{
A character vector of colors.
}
\description{
This function returns a palette designed for \code{densityPlot}.
}
\examples{
dpColors('sea')

}
