% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{labelPoints}
\alias{labelPoints}
\title{Label points in a ggplot object}
\usage{
labelPoints(
  p,
  labelDF,
  labelType = c("free", "boxed"),
  labelSize = 2,
  labelColor = "black",
  labelRepulsion = 1,
  labelPull = 1,
  maxOverlaps = Inf,
  ...
)
}
\arguments{
\item{p}{A ggplot object.}

\item{labelDF}{Label data frame.}

\item{labelType}{Whether to draw a box around labels (option 'boxed') or not
(option 'free'). Default is 'free'.}

\item{labelSize}{Label size.}

\item{labelColor}{Label color.}

\item{labelRepulsion}{Repulsion strength between labels.}

\item{labelPull}{Attraction strength between a text label
and its data point.}

\item{maxOverlaps}{Maximum overlaps.}

\item{...}{Additional arguments passed to \code{geom_text_repel}
(if \code{labelType} is 'free') or \code{geom_text_label}
(if \code{labelType} is 'boxed').}
}
\value{
A ggplot object.
}
\description{
This function labels points in a ggplot object.
}
