% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{labelPoints}
\alias{labelPoints}
\title{Label points in a ggplot object}
\usage{
labelPoints(
  p,
  labelDF,
  labelSize = 2,
  labelColor = "black",
  labelRepulsion = 1,
  labelPull = 1,
  maxOverlaps = Inf
)
}
\arguments{
\item{p}{A ggplot object.}

\item{labelDF}{Label data frame.}

\item{labelSize}{Label size.}

\item{labelColor}{Label color.}

\item{labelRepulsion}{Repulsion strength between labels.}

\item{labelPull}{Attraction strength between a text label
and its data point.}

\item{maxOverlaps}{Maximum overlaps.}
}
\value{
A ggplot object.
}
\description{
This function labels points in a ggplot object.
}
\keyword{internal}
