% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_plot.R
\name{networkPlotDF}
\alias{networkPlotDF}
\title{Prepare data frame for network plot}
\usage{
networkPlotDF(df, rankCol = "rank", edgeScale = 2)
}
\arguments{
\item{df}{Data frame.}

\item{rankCol}{Name of the rank column.}

\item{edgeScale}{Scaling factor used in generating
edge weights.}
}
\value{
A data frame ready to serve as input to \code{networkPlot}.
}
\description{
This function prepares a data frame for network plot.
}
\keyword{internal}
