% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_plot.R
\name{rankPlot}
\alias{rankPlot}
\title{Create a rank plot}
\usage{
rankPlot(
  df,
  title = "Rank plot",
  summarize = TRUE,
  viridisPal = "turbo",
  xLab = "Item",
  showMeanRanks = FALSE,
  labelSize = 3.1,
  labelColor = "seashell",
  ...
)
}
\arguments{
\item{df}{A data frame with ranks as columns and items as rows, or a
summary data frame generated with \code{rankSummary}. If the latter,
\code{summarize} must be set to \code{FALSE}.}

\item{title}{Plot title.}

\item{summarize}{Whether to summarize the ranks with \code{rankSummary}.
Must be set to \code{FALSE} if the input data frame has been generated with
\code{rankSummary}.}

\item{viridisPal}{Viridis palette.}

\item{xLab}{Label of x axis.}

\item{showMeanRanks}{Whether mean rank values should be displayed on
the plot.}

\item{labelSize}{Size of label marking average rank for each item. Ignored
if \code{showMeanRanks} is \code{FALSE}.}

\item{labelColor}{Color of label marking average rank for each item. Ignored
if \code{showMeanRanks} is \code{FALSE}.}

\item{...}{Other arguments passed to centerTitle.}
}
\value{
A ggplot object.
}
\description{
This function creates a rank plot
}
\examples{
df <- do.call(cbind, lapply(seq(30), function(i) sample(10, 10)))
rownames(df) <- paste0('M', seq(10))
colnames(df) <- paste0('R', seq(30))
rankPlot(df)

}
