% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_plot.R
\name{rankSummary}
\alias{rankSummary}
\title{Create a rank summary}
\usage{
rankSummary(df)
}
\arguments{
\item{df}{A data frame with ranks as columns and items as rows.}
}
\value{
A rank summary data frame with three columns: 'Rank', 'Item' and
'Count'.
}
\description{
This function creates a summary of multiple ranks provided for input items.
}
\examples{
df <- do.call(cbind, lapply(seq(30), function(i) sample(10, 10)))
rownames(df) <- paste0('M', seq(10))
colnames(df) <- paste0('R', seq(30))
rankSummary(df)

}
