% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_plot.R
\name{tilePlot}
\alias{tilePlot}
\title{Plot a numeric matrix or data frame}
\usage{
tilePlot(
  mat,
  title = "Tile plot",
  xLab = "x",
  yLab = "y",
  legendLab = "Value",
  isCor = FALSE,
  labelSize = 3,
  labelColor = "black",
  wesPalette = "Zissou1"
)
}
\arguments{
\item{mat}{A numeric matrix or data frame.}

\item{title}{Plot title.}

\item{xLab}{Label of x axis.}

\item{yLab}{Label of y axis.}

\item{legendLab}{Legend label.}

\item{isCor}{Whether the matrix is a correlation matrix, in which case the
limits of the color scale will be set to [-1, 1].}

\item{labelSize}{Label size. Ignored if \code{labelDF} is \code{NULL}.}

\item{labelColor}{labelColor}

\item{wesPalette}{wesanderson palette to be used for the color scale.}
}
\value{
A tile plot.
}
\description{
This function plots a numeric matrix or data frame.
}
\examples{
mat <- matrix(round(runif(100, 0, 1), 2), nrow=10)
rownames(mat) <- paste0('R', seq(10))
colnames(mat) <- paste0('C', seq(10))
tilePlot(mat)

}
