% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_plot.R
\name{tilePlot}
\alias{tilePlot}
\title{Plot a numeric matrix or data frame}
\usage{
tilePlot(
  mat,
  title = "Tile plot",
  xLab = "x",
  yLab = "y",
  legendLab = "Value",
  sigDigits = 2,
  isCor = FALSE,
  labelSize = 3,
  labelColor = "black",
  tileBoundaryColor = "white",
  tileBoundaryWidth = 0.2,
  palette = "Spectral",
  reverseColors = TRUE,
  xAngle = 45,
  vJust = 0.6,
  ...
)
}
\arguments{
\item{mat}{A numeric matrix or data frame.}

\item{title}{Plot title.}

\item{xLab}{Label of x axis.}

\item{yLab}{Label of y axis.}

\item{legendLab}{Legend label.}

\item{sigDigits}{Number of significant digits to be displayed for each
matrix element.}

\item{isCor}{Whether the matrix is a correlation matrix, in which case the
limits of the color scale will be set to [-1, 1].}

\item{labelSize}{Label size. Ignored if \code{labelDF} is \code{NULL}.}

\item{labelColor}{Label color.}

\item{tileBoundaryColor}{Tile boundary color.}

\item{tileBoundaryWidth}{Tile boundary width.}

\item{palette}{grDevices palette used for coloring nodes.
Ignored if nodeColor is not \code{NULL}.}

\item{reverseColors}{Whether to reverse the order of colors in the palette.}

\item{xAngle}{Angle of x axis text.}

\item{vJust}{Vertical justification in [0, 1].}

\item{...}{Other arguments passed to \code{centerTitle}.}
}
\value{
An object of class \code{gg}.
}
\description{
This function plots a numeric matrix or data frame.
}
\examples{
mat <- matrix(round(runif(100, 0, 1), 2), nrow=10)
rownames(mat) <- paste0('R', seq(10))
colnames(mat) <- paste0('C', seq(10))
tilePlot(mat)

}
